refactor(layout): deepen participant geometry
Concentrate participant ordering, sizing, alignment, and row placement behind one collaboration phase.
