+++ lib/layout/collaboration/index.js
@@ -1,14 +1,10 @@
 import { createLayout } from '../geometry/index.js';
 import {
-  orderParticipants
-} from './ordering/ParticipantOrdering.js';
-import {
-  positionParticipants
-} from './placement/ParticipantPlacement.js';
+  layoutParticipantGeometry
+} from './ParticipantGeometry.js';
 import {
   routeMessageFlows
 } from './routing/MessageFlowRouting.js';
-import { compactParticipantRows } from './steps/compactParticipantRows.js';
 import { layoutParticipants } from './steps/layoutParticipants.js';
 import { placeArtifacts } from './steps/placeArtifacts.js';
 import {
@@ -23,9 +19,7 @@ import {
 const COLLABORATION_LAYOUT_STEPS = Object.freeze([
   validateCollaboration,
   layoutParticipants,
-  orderParticipants,
-  positionParticipants,
-  compactParticipantRows,
+  layoutParticipantGeometry,
   routeMessageFlows,
   placeArtifacts
 ]);
